pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = ""
authors = ["DaniilMashkov <dankeeeez@gmail.com>"]
readme = "README.md"
packages = [{include = "blog_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
